requirements.txt
# Add Python dependencies here
semantic-kernel>=0.9.0
PyGithub>=2.0.
